Add getAll endpoint implementation
During implementation I faced an issue with generating hal paged response.
Spring's built in PagedResourcesAssembler class works with JPA entities
which means that the service class has to expose JPA types to the controller.
Another option would be to keep hal manipulations in the service layer,
but it would mean introducing operations that belong to formatting response
to the layer that should only handle business logic. I decided to sacrifice
JPA entities abstraction in order to keep service layer "clean".