Enhance parent-child hirerchy support
Add endpoint that lists subproducts for a given product. In addition
a product that is a child has also a link ref to the root of the tree.
With these changes client can traverse not only in one direction up,
but down to children or jump to the root of the hierarchy.